feature/UX-36:  * Move pipeline plugin code  * Refactor out a decorating component that talks to the store, apply to pipeline page  * Decorate TopNavPipelineCounts with pipelines helper, more clean up
Squashed commit of the following:

commit 6200affb8fd65db28096866dda4c153eb34e9eca
commit 601fa0f64002472574a3fb63b837ad54ab6fc2f3
commit 35a1f28e5e11541e810eed4e8da11852b5a9fb7c
commit 038c6dbe7ce2a280d0d3357824d1491106e8afe5
