+++ js-plugin-experiments/scripts/plugins/pipelines/index.jsx
@@ -1,0 +1,8 @@
+/**
+ * There will be a lot more here as the "bundle all the pipelines stuff as one plugin" process continues :D
+ */
+import PipelinesPage from './PipelinesPage';
+import TopNavPipelineCounts from './TopNavPipelineCounts';
+import MyPipelineRowExtension from './MyPipelineRowExtension';
+
+export {PipelinesPage, TopNavPipelineCounts, MyPipelineRowExtension};
